requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
Pillow
django-filter==23.5
qrcode
reportlab
boto3
django-storages
celery
redis
django-celery-beat
gunicorn
